Update TimeStepCopy.py
Added a function to delete the old time step. this function works for a single object, but if there are multiple the loop runs infinitely and crashes the program. I suspect this is because it copies one of the objects an extra time and adds it to the things it has to loop through.